Add more size_t

+++ include/al_dmabuf.h
@@ -23,12 +23,12 @@ struct al5_buffer_info {
 	u32 size;
 };
 
-void *al5_dmabuf_wrap(struct device *dev, unsigned long size,
+void *al5_dmabuf_wrap(struct device *dev, size_t size,
 		      struct al5_dma_buffer *buffer);
-int al5_create_dmabuf_fd(struct device *dev, unsigned long size,
+int al5_create_dmabuf_fd(struct device *dev, size_t size,
 			 struct al5_dma_buffer *buffer);
 
-int al5_allocate_dmabuf(struct device *dev, int size, u32 *fd);
+int al5_allocate_dmabuf(struct device *dev, size_t size, u32 *fd);
 int al5_dmabuf_get_address(struct device *dev, u32 fd, u32 *bus_address);
 int al5_get_dmabuf_info(struct device *dev, u32 fd,
 			struct al5_buffer_info *info);
